requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0

# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Time Series
prophet==1.1.4
statsmodels==0.14.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Framework
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.3

# Database
psycopg2-binary==2.9.6
sqlalchemy==2.0.19

# API & Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.8.5

# Geospatial
geopy==2.3.0
folium==0.14.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
joblib==1.3.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
ipywidgets==8.1.0

# Progress bars & logging
tqdm==4.65.0
loguru==0.7.0

# Data validation
great-expectations==0.17.11

# Model interpretability
shap==0.42.1
lime==0.2.0.1

# Streamlit (alternative frontend)
streamlit==1.25.0

# Date handling
python-dateutil==2.8.2
pytz==2023.3

# HTTP client
httpx==0.24.1